% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fib_under.R
\name{fib_under}
\alias{fib_under}
\title{Fibonacci checker fuction
Version 1.0
fib_under()
Finds the number of numbers in the Fibonacci sequence lower than or equal to the given input}
\usage{
fib_under(x)
}
\arguments{
\item{x}{Numeric input}
}
\value{
Number of Fibonacci numbers less than the input along with the numbers themselves
}
\description{
Fibonacci checker fuction
Version 1.0
fib_under()
Finds the number of numbers in the Fibonacci sequence lower than or equal to the given input
}
