% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fib.R, R/Fibonacci.R
\docType{package}
\name{Fibonacci}
\alias{Fibonacci}
\title{Fibonacci}
\usage{
Fibonacci(x)
}
\description{
This package is used to calculate Fibonacci number

Fibonacci number
}
\details{
Fibonacci @keywords
}
\section{Fibonacci Functions}{

Fibonacci Functions
}

\examples{
Fibonacci(10)
}
