% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci-data.R
\docType{data}
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci numbers}
\format{
An numerical vector [1:1000]
}
\usage{
data(fibonacci)
}
\description{
First 1000 Fibonacci numbers
}
\examples{
data(fibonacci)
}
\keyword{datasets}
